app/requirements.txt
arrow==1.3.0
asgiref==3.9.1
babel==2.17.0
binaryornot==0.4.4
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cookiecutter==2.6.0
crispy-tailwind==1.0.3
Django==5.2.4
django-browser-reload==1.18.0
django-crispy-forms==2.4
django-phonenumber-field==8.1.0
django-rest-framework==0.1.0
django-tailwind==4.2.0
djangorestframework==3.16.0
dotenv==0.9.9
gunicorn==23.0.0
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==25.0
phonenumbers==9.0.12
psycopg2-binary==2.9.10
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.4
rich==14.1.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
types-python-dateutil==2.9.0.20250708
urllib3==2.5.0
weasyprint
django_extensions
